ARC-105_search: restructure sidebar files
