+++ src/views/sidebar/components/main-search-tab/main-search-tab.scss
