--- src/views/sidebar/components/sidebar-summary/sidebar-summary.scss
@@ -1,20 +1,0 @@
-@import '../../../../defaults';
-
-.sidebar-summary {
-    width: 100%;
-    border-collapse: collapse;
-   
-    &-row {
-        height: 23px;
-        line-height: 23px;
-
-        &:nth-child(odd) {
-            background-color: $light-gray;
-        }
-    }
-    
-    &-label {
-        padding-left: 3px;
-        font-weight: bold;
-    }
-}
