+++ src/views/sidebar/components/summary-tab/summary-tab.scss
